+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/layout/FlowLayout.java
@@ -1,4 +1,5 @@
 package com.github.bordertech.wcomponents.layout;
+import com.github.bordertech.wcomponents.Size;
 import com.github.bordertech.wcomponents.WPanel;
 import com.github.bordertech.wcomponents.util.SpaceUtil;
 
@@ -91,7 +92,7 @@ public class FlowLayout implements LayoutManager {
 	/**
 	 * The space between components added to the FlowLayout. The direction of the space is determined by the Alignment.
 	 */
-	private final SpaceUtil.Size space;
+	private final Size space;
 
 	/**
 	 * For temporary backwards compatibility only.
@@ -112,7 +113,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param gap the requested space between components in the layout
 	 */
 	@Deprecated
-	private FlowLayout(final Alignment alignment, final SpaceUtil.Size space, final ContentAlignment contentAlignment, final int gap) {
+	private FlowLayout(final Alignment alignment, final Size space, final ContentAlignment contentAlignment, final int gap) {
 		if (alignment == null) {
 			throw new IllegalArgumentException("Alignment must be provided.");
 		}
@@ -157,7 +158,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param hgap The horizontal gap between the cells. Not used if alignment is VERTICAL.
 	 * @param vgap The vertical gap between the cells. Used only if alignment is VERTICAL.
 	 *
-	 * @deprecated use {@link #FlowLayout(Alignment, SpaceUtil.Size)}
+	 * @deprecated use {@link #FlowLayout(Alignment, Size)}
 	 */
 	@Deprecated
 	public FlowLayout(final Alignment alignment, final int hgap, final int vgap) {
@@ -171,7 +172,7 @@ public class FlowLayout implements LayoutManager {
 	 *
 	 * @param alignment the required alignment
 	 * @param gap the required gap between components in the layout
-	 * @deprecated use {@link #FlowLayout(Alignment, SpaceUtil.Size)}
+	 * @deprecated use {@link #FlowLayout(Alignment, Size)}
 	 */
 	@Deprecated
 	public FlowLayout(final Alignment alignment, final int gap) {
@@ -185,7 +186,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param alignment the required alignment
 	 * @param space the required space between components in the layout
 	 */
-	public FlowLayout(final Alignment alignment, final SpaceUtil.Size space) {
+	public FlowLayout(final Alignment alignment, final Size space) {
 		this(alignment, space, null);
 	}
 
@@ -199,7 +200,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param vgap The vertical gap between the cells. Used only if alignment is VERTICAL.
 	 * @param contentAlignment The relative vertical alignment of the content in each cell. Not used if alignment is VERTICAL.
 	 *
-	 * @deprecated use {@link #FlowLayout(Alignment, SpaceUtil.Size, ContentAlignment)}
+	 * @deprecated use {@link #FlowLayout(Alignment, Size, ContentAlignment)}
 	 */
 	@Deprecated
 	public FlowLayout(final Alignment alignment, final int hgap, final int vgap, final ContentAlignment contentAlignment) {
@@ -216,7 +217,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param alignment the required alignment
 	 * @param gap the required gap between components in the layout
 	 * @param contentAlignment the alignment of the content in each cell when alignment is not vertical
-	 * @deprecated use {@link #FlowLayout(Alignment, SpaceUtil.Size, ContentAlignment)}
+	 * @deprecated use {@link #FlowLayout(Alignment, Size, ContentAlignment)}
 	 */
 	@Deprecated
 	public FlowLayout(final Alignment alignment, final int gap, final ContentAlignment contentAlignment) {
@@ -232,7 +233,7 @@ public class FlowLayout implements LayoutManager {
 	 * @param space the required space between components in the layout
 	 * @param contentAlignment the alignment of the content in each cell when alignment is not vertical
 	 */
-	public FlowLayout(final Alignment alignment, final SpaceUtil.Size space, final ContentAlignment contentAlignment) {
+	public FlowLayout(final Alignment alignment, final Size space, final ContentAlignment contentAlignment) {
 		this(alignment, space, contentAlignment, -1);
 	}
 
@@ -264,7 +265,7 @@ public class FlowLayout implements LayoutManager {
 	/**
 	 * @return the space between the components added to the FlowLayout
 	 */
-	public SpaceUtil.Size getSpace() {
+	public Size getSpace() {
 		return space;
 	}