+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/render/webxml/ColumnLayoutRenderer.java
@@ -1,12 +1,12 @@
 package com.github.bordertech.wcomponents.render.webxml;
 
 import com.github.bordertech.wcomponents.Renderer;
+import com.github.bordertech.wcomponents.Size;
 import com.github.bordertech.wcomponents.WComponent;
 import com.github.bordertech.wcomponents.WPanel;
 import com.github.bordertech.wcomponents.XmlStringBuilder;
 import com.github.bordertech.wcomponents.layout.ColumnLayout;
 import com.github.bordertech.wcomponents.servlet.WebXmlRenderContext;
-import com.github.bordertech.wcomponents.util.SpaceUtil;
 
 /**
  * This {@link Renderer} renders the children of a {@link WPanel} which have been arranged using a {@link ColumnLayout}.
@@ -29,9 +29,9 @@ final class ColumnLayoutRenderer extends AbstractWebXmlRenderer {
 		XmlStringBuilder xml = renderContext.getWriter();
 		ColumnLayout layout = (ColumnLayout) panel.getLayout();
 		int childCount = panel.getChildCount();
-		SpaceUtil.Size hgap = layout.getHorizontalGap();
+		Size hgap = layout.getHorizontalGap();
 		String hgapString = hgap == null ? null : hgap.toString();
-		SpaceUtil.Size vgap = layout.getVerticalGap();
+		Size vgap = layout.getVerticalGap();
 		String vgapString = vgap == null ? null : vgap.toString();
 		int cols = layout.getColumnCount();