+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/render/webxml/WListRenderer_Test.java
@@ -1,6 +1,7 @@
 package com.github.bordertech.wcomponents.render.webxml;
 
 import com.github.bordertech.wcomponents.Margin;
+import com.github.bordertech.wcomponents.Size;
 import com.github.bordertech.wcomponents.WList;
 import com.github.bordertech.wcomponents.WText;
 import com.github.bordertech.wcomponents.util.SpaceUtil;
@@ -23,12 +24,12 @@ public class WListRenderer_Test extends AbstractWebXmlRendererTestCase {
 	/**
 	 * A reusable gap value.
 	 */
-	private static final SpaceUtil.Size GAP = SpaceUtil.Size.SMALL;
+	private static final Size GAP = Size.SMALL;
 
 	/**
 	 * A different reusable gap value. This is used to differentiate the (now deprecated) hgap and vgap properties.
 	 */
-	private static final SpaceUtil.Size BIG_GAP = SpaceUtil.Size.LARGE;
+	private static final Size BIG_GAP = Size.LARGE;
 
 	@Test
 	public void testLayoutCorrectlyConfigured() {
@@ -127,7 +128,7 @@ public class WListRenderer_Test extends AbstractWebXmlRendererTestCase {
 		list.setMargin(margin);
 		assertXpathNotExists("//ui:panel/ui:margin", list);
 
-		margin = new Margin(SpaceUtil.Size.SMALL);
+		margin = new Margin(Size.SMALL);
 		list.setMargin(margin);
 		assertSchemaMatch(list);
 		assertXpathEvaluatesTo("sm", "//ui:panel/ui:margin/@all", list);
@@ -136,7 +137,7 @@ public class WListRenderer_Test extends AbstractWebXmlRendererTestCase {
 		assertXpathEvaluatesTo("", "//ui:panel/ui:margin/@south", list);
 		assertXpathEvaluatesTo("", "//ui:panel/ui:margin/@west", list);
 
-		margin = new Margin(SpaceUtil.Size.SMALL, SpaceUtil.Size.MEDIUM, SpaceUtil.Size.LARGE, SpaceUtil.Size.XL);
+		margin = new Margin(Size.SMALL, Size.MEDIUM, Size.LARGE, Size.XL);
 		list.setMargin(margin);
 		assertSchemaMatch(list);
 		assertXpathEvaluatesTo("", "//ui:panel/ui:margin/@all", list);