+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/theme/WPanelTypeExample.java
@@ -5,6 +5,7 @@ import com.github.bordertech.wcomponents.ActionEvent;
 import com.github.bordertech.wcomponents.HeadingLevel;
 import com.github.bordertech.wcomponents.Margin;
 import com.github.bordertech.wcomponents.Request;
+import com.github.bordertech.wcomponents.Size;
 import com.github.bordertech.wcomponents.WAjaxControl;
 import com.github.bordertech.wcomponents.WButton;
 import com.github.bordertech.wcomponents.WCheckBox;
@@ -28,7 +29,6 @@ import com.github.bordertech.wcomponents.subordinate.Or;
 import com.github.bordertech.wcomponents.subordinate.Rule;
 import com.github.bordertech.wcomponents.subordinate.Show;
 import com.github.bordertech.wcomponents.subordinate.WSubordinateControl;
-import com.github.bordertech.wcomponents.util.SpaceUtil;
 import com.github.bordertech.wcomponents.util.HtmlClassProperties;
 import com.github.bordertech.wcomponents.util.Util;
 
@@ -236,7 +236,7 @@ public class WPanelTypeExample extends WContainer {
 	 */
 	private void buildConfigOptions() {
 		WFieldLayout layout = new WFieldLayout(WFieldLayout.LAYOUT_STACKED);
-		layout.setMargin(new Margin(null, null, SpaceUtil.Size.LARGE, null));
+		layout.setMargin(new Margin(null, null, Size.LARGE, null));
 		layout.addField("Select a WPanel Type", panelType);
 		contentField = layout.addField("Panel content", panelContent);
 		headingField = layout.addField("heading", tfHeading);