+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/validation/repeater/RepeaterComponent.java
@@ -1,12 +1,12 @@
 package com.github.bordertech.wcomponents.examples.validation.repeater;
 
 import com.github.bordertech.wcomponents.Margin;
+import com.github.bordertech.wcomponents.Size;
 import com.github.bordertech.wcomponents.WDataRenderer;
 import com.github.bordertech.wcomponents.WField;
 import com.github.bordertech.wcomponents.WFieldLayout;
 import com.github.bordertech.wcomponents.WFieldSet;
 import com.github.bordertech.wcomponents.WTextField;
-import com.github.bordertech.wcomponents.util.SpaceUtil;
 import com.github.bordertech.wcomponents.validator.RegExFieldValidator;
 
 /**
@@ -36,7 +36,7 @@ public class RepeaterComponent extends WDataRenderer {
 	public RepeaterComponent() {
 		fieldSet = new WFieldSet("Repeated Fields");
 		add(fieldSet);
-		fieldSet.setMargin(new Margin(null, null, SpaceUtil.Size.LARGE, null));
+		fieldSet.setMargin(new Margin(null, null, Size.LARGE, null));
 		WFieldLayout fields = new WFieldLayout();
 		fieldSet.add(fields);